优化我喜欢的音乐

+++ app/src/main/java/me/wcy/music/common/bean/PlaylistData.kt
@@ -17,7 +17,7 @@ data class PlaylistData(
     @SerializedName("trackCount")
     val trackCount: Int = 0,
     @SerializedName("userId")
-    val userId: Int = 0,
+    val userId: Long = 0,
     @SerializedName("playCount", alternate = ["playcount"])
     val playCount: Long = 0,
     @SerializedName("bookCount")
